# Initial support for unity builds - speeds up compilation by up to 58%