CMake 3.16 added support for unity builds and this PR introduces them to redpanda. For a detailed description of unity builds along with the pros/cons check out [this article](https://onqtam.com/programming/2019-12-20-pch-unity-cmake-3-16/).

The measurements were done in the following way:
- Ubuntu 19.10, GCC 9.2.1, CMake 3.19, i7-8750H CPU @ 2.20GHz × 12, 32GB RAM, 1TB SSD
- 2 scenarios were tested - with up to 10 sources in a unity batch (so a target with 26 sources will end up with 3 unity files with 10/10/6 accordingly) and bundling all sources of a target into a single unity file
- measurements were taken with `time ninja` using the `real` time - CMake configuration and generation time is not included - only rebuilding whatever is left after a full build followed a clean (`ninja -t clean`)
- the first row is for a full clean build (after `ninja -t clean`) and all other rows are for minimal rebuilds by touching the respective file

|                     | baseline  | unity x10 | unity x10<br>gain % | unity x10<br>gain sec | unity all | unity all<br>gain % | unity all<br>gain sec |
|---------------------|-----------|-----------|------------------------|--------------------------|------------|-------------------------|---------------------------|
| **FULL BUILD**      | 12m5,620s | 5m57,100s | +51%                   | +368 sec                 | 5m17,877s  | +56%                    | +407 sec                  |
| `record.h`          | 10m3,643s | 4m46,272s | +52%                   | +317 sec                 | 4m15,612s  | +58%                    | +348 sec                  |
| `segment.h`         | 7m23,491s | 4m1,521s  | +45%                   | +202 sec                 | 3m25,835s  | +53%                    | +237 sec                  |
| `consensus.h`       | 5m44,449s | 2m40,766s | +53%                   | +184 sec                 | 2m7,460s   | +63%                    | +217 sec                  |
| `partition.h`       | 3m49,769s | 1m54,835s | +50%                   | +115 sec                 | 1m33,474s  | +60%                    | +137 sec                  |
| `group_manager.h`   | 3m47,872s | 2m5,896s  | +45%                   | +102 sec                 | 1m47,395s  | +53%                    | +121 sec                  |
| `kafka/client.h`    | 1m51,571s | 1m0,008s  | +46%                   | +51 sec                  | 1m8,101s   | +39%                    | +43 sec                   |
| `application.cc`    | 0m48,131s | 0m43,289s | +11%                   | +5 sec                   | 0m45,653s  | +5%                     | +2 sec                    |
| `segment.cc`        | 0m41,009s | 0m39,743s | +3%                    | +1 sec                   | 0m40,141s  | +3%                     | +1 sec                    |
| `consensus.cc`      | 0m40,537s | 0m39,641s | +2%                    | +1 sec                   | 0m39,795s  | +2%                     | +1 sec                    |
| `raft/types.cc`     | 0m34,009s | 0m40,394s | -17%                   | -6 sec                   | 0m47,479s  | -38%                    | -13 sec                   |
| `segment_reader.cc` | 0m28,268s | 0m39,084s | -39%                   | -11 sec                  | 0m59,345s  | -110%                   | -31 sec                   |
| `fetch_request.cc`  | 0m24,892s | 0m29,073s | -16%                   | -4 sec                   | 0m46,349s  | -84%                    | -21 sec                   |
| `fetch_test.cc`     | 0m24,540s | 0m39,571s | -62%                   | -15 sec                  | 0m45,652s  | -87%                    | -21 sec                   |
| `cluster_utils.cc`  | 0m19,037s | 0m32,086s | -68%                   | -13 sec                  | 0m56,767s  | -194%                   | -37 sec                   |

As can be seen in the table touching single source files might be slower in a unity build and the negative effect is more pronounced in bigger targets when all sources are clumped together. In this PR source files are batched together in groups of 10 as the better compromise but "*better*" is arguable because usually along with changes to a source file come some changes in a header file as well so perhaps batching everything in a single unity file could be worth it.

## Further work:
- stop relying on the installed version of CMake on the host system and bundle CMake 3.17.2 (or newer) in the build
- merge most test binaries into fewer ones so that unity builds can be useful (especially for storage/utils/kafka/model/etc. which have a bunch of tests in a single `CMakeLists.txt`) - there can be separate CMake commands to register separate ctests if necessary - this is a big opportunity to further cut down build times by at least 1 minute

## Other notes
- I had to install a newer version of hwloc and also copy `xxhash.c` right next to `/usr/include/xxhash.h` in order to build the project - haven't investigated in depth.
- The unity build does not consume more RAM than normal builds (or perhaps a slight increase of about 10%)
- Could produce faster binaries because unity builds are effectively the same as `-flto` (building the third-party libs as unity might be beneficial as well)
- This patch should play well with `ccache` but I haven't tested it.
- I tried using precompiled headers for the cluster/raft/kafka/storage targets and experimented with both `model/record.h` which ends up almost everywhere in `src/v/` (it produced a PCH of 300MB+ for each target) and just `<seastar/core/future.hh>` (smaller - about 80MB) but in both cases, the effects were either negligible or even detrimental - this codebase gains 0 from saving on parsing because it's dominated by template instantiation
- Overall I expected better results but the targets have very few source files and dependencies are as granular as possible and thus there isn't that much to be gained from link time with smaller object files. This codebase is of exceptional quality.

## Checklist
- [ ] Reference related [issue](https://github.com/vectorizedio/redpanda/issues)
- [ ] Update [PendingReleaseNotes.md](https://github.com/dotnwat/redpanda/blob/dev/PendingReleaseNotes.md), if relevant