@onqtam 2020-11-15
trying with CMake 3.16 instead of 3.17.2 as the minimum required to see if the builds pass