@BenPope 2020-11-15
Hi, thanks for this. I have previously had a quick play with unity builds. It seemed to be not very effective when `ccache` is in use.  

After this I suspect that link time starts to dominate, are you using `ld.gold`?

I also looked briefly at PCH, but it appears that it generates too often, especially around tests.

`iobuf.h` is the worst offender for compile time; it includes a lot and ends up everywhere. I've used https://github.com/aras-p/ClangBuildAnalyzer