@onqtam 2020-11-15
@BenPope not using gold - just ld - I could play with gold as well. About PCH - even if it gets recompiled only when necessary it still delivers 0 value for this codebase.