@BenPope 2020-11-15
Clang 11 has gained `-fpch-instantiate-templates` but I haven't tried it.