@onqtam 2020-11-15
> Clang 11 has gained `-fpch-instantiate-templates` but I haven't tried it.

Cool - didn't know about that :)

Regarding ccache and unity builds - do you remember what the problems were? ccache should be useful for full builds with minimal changes but when impactful header files were touched (such as `iobuf.h` or anything dragged by it) then unity builds would provide compounding value.