@BenPope 2020-11-15
I didn't play with unity builds extensively, at the time we only had about 55% ccache hit rates due to a compile random number generation, which I've removed. ccache should have 100% hit rate now. I don't recall any issues with unity builds and ccache, but I could imagine if they don't chunk up the files consistently, ccache would miss.

`ccache` and `ld.gold` are highly recommended.