@BenPope 2020-11-15
Yes, that code is under heavy development. It should be fixed soon.