@senior7515 2020-11-15
> Are there sporadic failures in ctest? 1 failed on the build: https://github.com/vectorizedio/redpanda/pull/91/checks?check_run_id=1403296370#step:6:17039
> 
> ```
> The following tests FAILED:
> 	 37 - read_materialized_topic_unit_tests_rpunit (Failed)
> ```

@onqtam thanks for this patch. @mmaslankaprv has a fix for it here https://github.com/vectorizedio/redpanda/pull/89

Waiting for v2 review. Should be fixed tom.