@dotnwat 2020-11-16
Thanks for submitting this it is really cool. Especially if this speeds up CI (full builds) this will be great. As ben mentioned, optimizing for developer experience is also really important, and for that ccache is critical (as well as our never ending process of minimizing header dependences). Perhaps we could even turn on unity builds in CI if ccache issues popped up for normal development experience?