@BenPope 2020-11-16
This does make a lot of sense for CI and one-time builds. My concern is introducing duplicate symbols and breaking CI without it breaking locally, but that's likely a bug anyway which we don't spot at all, and there doesn't seem to be too many at the moment. It may take a little while to settle, as the set of chunks moves around. Perhaps we can find a static analysis tool for spotting duplicate symbols ahead of time.