@dotnwat 2020-11-16
good point @BenPope, sounds like we simply need an easy mechanism for turning this on and off. this PR seems to show that after all this time we've only managed to introduce a couple duplicate symbols. as long as we can reproduce locally, it seems like duplicate symbols are a pretty low risk issue.