@onqtam 2020-11-17
I reworked the PR - here are the changes:
- No longer requiring a specific CMake version for unity builds - turns out all unity-related CMake code is just setting properties or global `CMAKE_*` variables and if there is no unity build support (say... version 3.15) then it will simply not be built as unity. Thus there was no need to surround any CMake code with checks. I tested this with CMake 3.15.
- Unity builds are disabled by default but now they can be enabled by `./build.sh -DCMAKE_UNITY_BUILD=ON` - that would get cached so there's no need to pass it every time - just when you are switching between `ON` and `OFF`. They are enabled on the GitHub Actions workflow.
- The coloured output has been factored into a separate PR: #114
- I decided that this still makes sense to be 1 single commit.

I just saw your concern from a few days ago about memory usage and gcc reaching up to 3gb for a single translation unit - I looked at the overall ram usage for a full build with 12 threads with ninja and concluded that it was almost the same as the non-unity build (or within 10% margin).

I also tested this with ccache and the 2 work together properly.

My opinion is that it makes even more sense to use unity builds on local machines during development and ccache on the CI (by somehow reusing the cache between builds as artifacts). My understanding is that ccache caches full translation units after the preprocessor and maps those to object files. So if I'm working locally and I pull changes which end up affectiong say... `record.h` then I would still need to rebuild almost the entire codebase (ninja will rebuild all object files where that header ends up by transitive inclusion) - ccache doesn't help with that but a unity build does. I tried `touch`-ing that header and then ccache works proeprly because a touch is not a modification ==> same contents after the preprocessor, but if I added a single empty line then almost everything got rebuilt. It makes sense to me that ccache does not help with minimal rebuilds, and in a codebase with proper dependencies minimal rebuilds are the only thing I would be doing - even when switching branches.

In my previous company we enabled unity builds everywhere - both for CI and for local workflows - the only downside to that is that you might forget to `#include` a header in some `.cc` file because it gets dragged from another `.cc` above in the unity file, and for that we had a separate CI job which built the codebase without unity, and thus we had to occasionally add a header or two.

Anyway, I don't mean to push too hard for this and perhaps I might be misunderstanding something, but you should give it a try locally with `./build.sh -DCMAKE_UNITY_BUILD=ON`.

Let me know if this is good enough.