@onqtam 2020-11-18
@BenPope Just force-pushed the changes into 3 separate commits.

Regarding `compile_commands.json` - there's a [bug](https://gitlab.kitware.com/cmake/cmake/-/issues/19826) reported in the CMake issue tracker. I don't think this would be too hard to fix on the CMake side and am wondering if I should take a stab at it. There is also one 1 ugly workaround for that: run CMake twice - once with `CMAKE_EXPORT_COMPILE_COMMANDS=ON` and `CMAKE_UNITY_BUILD=OFF` and then once more with `CMAKE_EXPORT_COMPILE_COMMANDS=OFF` and `CMAKE_UNITY_BUILD=ON`, but that could open another can of worms.